Move: Point 도메인 분리를 위해, 애플리케이션 계층 이동
