+++ src/test/kotlin/io/ticketaka/api/payment/application/PaymentServiceTest.kt
@@ -1,7 +1,7 @@
 package io.ticketaka.api.payment.application
 
-import io.ticketaka.api.reservation.application.PaymentService
-import io.ticketaka.api.reservation.application.dto.PaymentCommand
+import io.ticketaka.api.point.application.PaymentService
+import io.ticketaka.api.point.application.dto.PaymentCommand
 import io.ticketaka.api.reservation.domain.payment.Payment
 import io.ticketaka.api.reservation.domain.payment.PaymentRepository
 import io.ticketaka.api.reservation.domain.point.Point
